refactor: migrate authentication context to Zustand store and update related components
