+++ frontend/src/components/ProtectedRoute.jsx
@@ -1,8 +1,9 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
+import useAuthStore from '../store/useAuthStore';
 
 const ProtectedRoute = ({ children, adminOnly = false }) => {
-  const { user, loading } = useAuth();
+  
+  const { user, loading } = useAuthStore();
 
   if (loading) {
     return <div>Loading...</div>;
