--- frontend/src/context/AuthContext.jsx
@@ -1,44 +1,0 @@
-import { createContext, useState, useEffect, useContext } from 'react';
-import { getMe } from '../api';
-
-const AuthContext = createContext();
-
-export const AuthProvider = ({ children }) => {
-  const [user, setUser] = useState(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const checkUser = async () => {
-      const token = localStorage.getItem('token');
-      if (token) {
-        try {
-          const { data } = await getMe();
-          setUser(data);
-        } catch (error) {
-          console.error('Auth check failed', error);
-          localStorage.removeItem('token');
-          setUser(null);
-        }
-      }
-      setLoading(false);
-    };
-    checkUser();
-  }, []);
-
-  const login = (userData) => {
-    setUser(userData);
-  };
-  
-  const logout = () => {
-    localStorage.removeItem('token');
-    setUser(null);
-  };
-
-  return (
-    <AuthContext.Provider value={{ user, login, logout, loading }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => useContext(AuthContext);
