+++ frontend/src/pages/DoctorDashboard.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
+import useAuthStore from '../store/useAuthStore';
 
 const StatCard = ({ value, label, currency = '' }) => (
     <div className="bg-secondary-dark p-6 rounded-xl text-center shadow-lg">
@@ -10,7 +10,7 @@ const StatCard = ({ value, label, currency = '' }) => (
 );
 
 const DoctorDashboard = () => {
-    const { user } = useAuth();
+    const { user } = useAuthStore();
     const role = user?.role || 'professional'; // 'doctor' or 'nurse'
 
     return (
