+++ frontend/src/pages/PatientDashboard.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
+import useAuthStore from '../store/useAuthStore';
 
 // --- New Reusable Feature Card Component ---
 // This new card has a gradient background and a cleaner look.
@@ -17,7 +17,7 @@ const NewFeatureCard = ({ icon, title, description, link, color }) => (
 
 
 const PatientDashboard = () => {
-  const { user } = useAuth();
+  const { user } = useAuthStore();
 
   return (
     // The background is now a clean, dark slate color from our tailwind config
